requirements.txt
PyMuPDF
pytesseract
Pillow
python-docx
faiss-cpu
torch
transformers
streamlit